requirements.txt
abcplus==0.1.0
alembic==1.6.5
appdirs==1.4.4
astroid
Babel==2.9.1
bcrypt==3.2.0
blinker==1.4
brotlipy==0.7.0
certifi==2021.5.30
cffi
charset-normalizer
click
colorama
cryptography==3.4.8
cssselect==1.1.0
dnspython==2.1.0
email-validator==1.1.3
fake-useragent==0.1.11
Flask
Flask-Caching==1.10.1
Flask-DebugToolbar==0.11.0
Flask-Login==0.5.0
Flask-Mail==0.9.1
Flask-SQLAlchemy
Flask-WTF==0.15.1
greenlet
gunicorn==20.1.0
idna
importlib-metadata
isort
itsdangerous
Jinja2
lazy-object-proxy
loguru==0.5.3
Mako==1.1.4
MarkupSafe
mccabe==0.6.1
mypy
mypy-extensions==0.4.3
parse==1.19.0
passlib==1.7.4
Pillow==8.3.2
psutil
psycopg2
psycopg2-binary==2.9.1
pycparser
pyee==8.2.2
pylint
pyOpenSSL
pyppeteer==0.2.6
pyquery==1.4.3
PySocks
python-dateutil==2.8.2
python-editor==1.0.4
pytz==2021.1
requests
requests-html==0.10.0
six
soupsieve==2.2.1
speaklater==1.3
SQLAlchemy
toml
tqdm==4.62.3
typing-extensions
w3lib==1.22.0
websockets==9.1
Werkzeug
win-inet-pton
win32-setctime==1.0.3
wincertstore==0.2
wrapt==1.12.1
WTForms==2.3.3
zipp
